backend/requirements.txt
fastapi>=0.100,<0.115
uvicorn[standard]>=0.22
sqlalchemy>=2.0
pydantic>=2.0
email-validator>=2.0
bcrypt>=4.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
